# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Echo Registry API",
    "description": "Latest Forge, NeoForge, Fabric, and popular mod versions for Minecraft",
    "version": "1.0.0",
    "contact": {
      "name": "Echo Registry"
    }
  },
  "servers": [
    {
      "url": "https://echo.iamkaf.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health check",
        "description": "Check the health status of the API and external services",
        "tags": ["Health"],
        "responses": {
          "200": {
            "description": "Health status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable"
          }
        }
      }
    },
    "/api/versions/minecraft": {
      "get": {
        "summary": "Get all Minecraft versions",
        "description": "Retrieve a list of all available Minecraft versions",
        "tags": ["Versions"],
        "responses": {
          "200": {
            "description": "List of Minecraft versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinecraftVersionsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch versions"
          }
        }
      }
    },
    "/api/versions/dependencies/{mcVersion}": {
      "get": {
        "summary": "Get all dependencies for Minecraft version",
        "description": "Retrieve all dependency versions for a specific Minecraft version",
        "tags": ["Dependencies"],
        "parameters": [
          {
            "name": "mcVersion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1.21.1"
          },
          {
            "name": "projects",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of additional Modrinth projects to include",
            "example": "fabric-api,modmenu,rei"
          }
        ],
        "responses": {
          "200": {
            "description": "List of dependency versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependenciesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch dependencies"
          }
        }
      }
    },
    "/api/projects/compatibility": {
      "get": {
        "summary": "Check project compatibility across loaders",
        "description": "Check which Modrinth projects are compatible with specific loaders across multiple Minecraft versions",
        "tags": ["Projects"],
        "parameters": [
          {
            "name": "projects",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of Modrinth project slugs",
            "example": "fabric-api,amber,sodium"
          },
          {
            "name": "versions",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of Minecraft versions",
            "example": "1.20.1,1.21.1"
          }
        ],
        "responses": {
          "200": {
            "description": "Project compatibility data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCompatibilityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Failed to fetch project data"
          }
        }
      }
    },
    "/api/cleanup": {
      "post": {
        "summary": "Clean up expired cache entries",
        "description": "Remove expired entries from the cache (for cron jobs)",
        "tags": ["Admin"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cleanup completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    },
                    "stats": {
                      "$ref": "#/components/schemas/CacheStats"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Cleanup failed"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DependencyVersion": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Dependency name"
          },
          "loader": {
            "type": "string",
            "enum": ["forge", "neoforge", "fabric", "universal"],
            "description": "Loader type"
          },
          "version": {
            "type": "string",
            "description": "Version string"
          },
          "mc_version": {
            "type": "string",
            "description": "Minecraft version"
          },
          "source_url": {
            "type": "string",
            "description": "Source URL for version information"
          },
          "download_urls": {
            "type": "object",
            "description": "Download URLs by mod loader (for Modrinth projects)",
            "properties": {
              "forge": {
                "type": ["string", "null"],
                "format": "uri",
                "description": "Download URL for Forge version"
              },
              "neoforge": {
                "type": ["string", "null"],
                "format": "uri",
                "description": "Download URL for NeoForge version"
              },
              "fabric": {
                "type": ["string", "null"],
                "format": "uri",
                "description": "Download URL for Fabric version"
              }
            }
          },
          "notes": {
            "type": "string",
            "description": "Additional notes about the version"
          },
          "fallback_used": {
            "type": "boolean",
            "description": "Whether a fallback version was used"
          },
          "cached_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this version was cached"
          }
        }
      },
      "MinecraftVersion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Version ID"
          },
          "version_type": {
            "type": "string",
            "enum": ["release", "snapshot", "old_beta", "old_alpha"],
            "description": "Version type"
          },
          "release_time": {
            "type": "string",
            "format": "date-time",
            "description": "Release timestamp"
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "required": ["success", "timestamp"],
        "properties": {
          "data": {
            "description": "Response data"
          },
          "error": {
            "type": "string",
            "description": "Error message (if any)"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Response timestamp"
          },
          "cached_at": {
            "type": "string",
            "format": "date-time",
            "description": "Cache timestamp"
          },
          "success": {
            "type": "boolean",
            "description": "Whether the request was successful"
          }
        }
      },
      "HealthResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": ["ok", "degraded", "down"]
              },
              "cache_status": {
                "type": "string",
                "enum": ["connected", "disconnected", "error"]
              },
              "external_apis": {
                "type": "object",
                "additionalProperties": {
                  "type": "string",
                  "enum": ["ok", "error"]
                }
              }
            },
            "required": ["status", "cache_status", "external_apis"]
          }
        ]
      },
      "DependenciesResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "mc_version": {
                    "type": "string"
                  },
                  "dependencies": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DependencyVersion"
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "MinecraftVersionsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "versions": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MinecraftVersion"
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "ProjectCompatibilityResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "description": "Nested mapping of projects to versions to loaders that support each project",
                "additionalProperties": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": ["forge", "neoforge", "fabric"]
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "CacheStats": {
        "type": "object",
        "properties": {
          "dependency_cache_count": {
            "type": "integer"
          },
          "minecraft_versions_count": {
            "type": "integer"
          },
          "expired_entries": {
            "type": "integer"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "Health",
      "description": "Health check endpoints"
    },
    {
      "name": "Versions",
      "description": "Minecraft version endpoints"
    },
    {
      "name": "Dependencies",
      "description": "Dependency version endpoints"
    },
    {
      "name": "Projects",
      "description": "Project compatibility endpoints"
    },
    {
      "name": "Admin",
      "description": "Administrative endpoints"
    }
  ]
}